requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg[binary]==3.1.12
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1
